Move json.hpp location - Move from astra-sim/ directory into extern/helper/json
