+++ astra-sim/network_frontend/analytical/congestion_aware/main.cc
@@ -8,7 +8,7 @@ LICENSE file in the root directory of this source tree.
 #include "AstraCongestionApi.hh"
 #include "astra-sim/system/Sys.hh"
 // #include "astra-sim/system/memory/SimpleMemory.hh"
-#include "astra-sim/json.hpp"
+#include <json/json.hpp>
 #include "event-queue/EventQueue.hh"
 #include "extern/memory_backend/analytical/AnalyticalMemory.hh" // Divya: to port congestion backend to latest Chakra
 #include "helper/CommandLineOptParser.hh"
