+++ astra-sim/network_frontend/analytical/congestion_unaware/main.cc
@@ -6,7 +6,7 @@ LICENSE file in the root directory of this source tree.
 #include <iostream>
 
 #include "AnalyticalNetwork.hh"
-#include "astra-sim/json.hpp"
+#include <json/json.hpp>
 #include "astra-sim/system/Sys.hh"
 #include "event-queue/EventQueue.hh"
 #include "event-queue/EventQueueEntry.hh"
